pyproject.toml
[tool.poetry]
name = "dailynews"
version = "0.1.0"
description = ""
authors = ["kevin"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
pymongo = "^4.11.3"
requests = "^2.32.3"
redis = "^5.2.1"
python-dateutil = "^2.9.0.post0"
scikit-learn = "^1.6.1"
flask = "^3.1.0"
kafka-python = "^2.0.2"
black = "^25.1.0"
newspaper3k = "^0.2.8"
confluent-kafka = "^2.9.0"
lxml-html-clean = "^0.4.1"
jsonrpcserver = "^5.0.9"
jsonrpcclient = "^4.0.3"
openai = "^1.70.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
